+++ fork/dynamic-cdn-webpack-plugin/src/find.test.js
@@ -17,12 +17,9 @@ describe('findPackages', () => {
 		'/node_modules/react/index.js': 'console.log("react");',
 		'/node_modules/react/package.json': '{"name": "react", "version": "16.14.0"}',
 		'/node_modules/@types/classnames/index.js': 'console.log("@types/classnames");',
-		'/node_modules/@talend/react-components/index.js':
-			'console.log("@talend/react-components");',
-		'/node_modules/@talend/react-components/node_modules/react/index.js':
-			'console.log("react");',
-		'/node_modules/@talend/react-containers/index.js':
-			'console.log("@talend/react-containers");',
+		'/node_modules/@talend/react-components/index.js': 'console.log("@talend/react-components");',
+		'/node_modules/@talend/react-components/node_modules/react/index.js': 'console.log("react");',
+		'/node_modules/@talend/react-containers/index.js': 'console.log("@talend/react-containers");',
 	};
 
 	beforeEach(() => {
@@ -30,11 +27,12 @@ describe('findPackages', () => {
 		require('fs').__setMockFiles(MOCK_FILE_INFO);
 	});
 
-	test('should find root and nested package', () => {
+	test.only('should find root and nested package', () => {
 		// when
 		const result = findPackages(undefined, 'react');
 
 		// then
+		console.log('LIST OF RESULT: ', result);
 		expect(result.length).toBe(2);
 		expect(result[0]).toBe('/node_modules/react');
 		expect(result[1]).toBe('/node_modules/@talend/react-components/node_modules/react');