+++ packages/design-tokens/jest.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	testRegex: '(/__tests__/.*|src/|scripts/).*\\.test.(js|ts|tsx)$',
+};