+++ tools/scripts-config-cdn/cdn.js
@@ -169,6 +169,9 @@ function getModulesFromLockFile(dir) {
 		const json = yaml.load(fs.readFileSync(lockTypeMap.pnpm.path, 'utf-8'));
 		infos = Object.keys(json.packages)
 			.map(moduleAndversion => {
+				if (moduleAndversion.startsWith('file:')) {
+					return null;
+				}
 				const moduleName = getModuleName(moduleAndversion, true);
 				return moduleToCdn(moduleName, getModuleVersion(moduleAndversion, true), {
 					env: 'development',